requirements.txt
genai>=0.11.0
python-dotenv>=1.0.0
streamlit>=1.28.0
markdown>=3.5.0
